Update patch set 1
Patch Set 1:

Note that the default is to launch a job with:

 LAUNCHER_COMMAND = 'hudson.slaves.CommandLauncher'


Whereas the code in Nodepool uses JNLP Launcher.

Whenever this patch is merged and released, Nodepool will be able to depend on the new python-jenkins version but will have to explicitly set the launcher to LAUNCHER_JNLP (or add support for different launchers types).

Patch-set: 1