Update patch set 1
Patch Set 1:

I would like to suggest one minor addition -- remove the .pep8 file. When it is present, flake8 seems to ignore the [flake8] section in tox.ini and as a result of that it scans the .tox directory and finds tons and tons of problems in various external packages. Removing the .pep8 file fixes this.

    $ tox -e pep8 | grep '^\./\.tox' | wc -l
      112526

    $ git rm .pep8
    rm '.pep8'

    $ tox -e pep8 | grep '^\./\.tox' | wc -l
           0

Patch-set: 1