Update patch set 8
Patch Set 8:

(7 comments)

Thanks, Victor! I'm going to upload a patch set that addresses at least one of your comments. Some of the other comments are very good, but I'm inclined to address those later, so as not to create merge conflicts for myself with the Python 3 PR.

Patch-set: 8
Label: Verified=0