Update patch set 22
Patch Set 22:

(1 comment)

Victor. Yep it works on py26 because I added unittest2 to testenv:py26 in tox.ini.

I could put it in test-requirements-py2.txt and reference that from tox.ini but it seemed like maybe overkill for one package. But I'll do if that's the standard or what folks prefer.

Patch-set: 22
Label: Verified=0