Update patch set 10
Patch Set 10:

Interesting point, Mathieu. I also debated pbr because python-jenkins is a very general module, potentially used by many outside OpenStack. And it is small as you said.

Also one thing that causes problems sometimes with pbr is the fact that it makes you use `setup_requires` and that is a problematic feature of setuptools in my experience. For example, see: https://bitbucket.org/pypa/setuptools/issue/209/setup_requires-and-install_requires-dont

Matthieu, would you like to see a change that adds pep8 checks and a "doc" tox target, but without pbr? And are others on board with that too?

Patch-set: 10