Update patch set 10
Patch Set 10:

Victor kindly provided a list of all packages done by upstream. python-pbr is available since Ubuntu Saucy. However, note that python-jenkins is also packaged for Ubuntu Precise which does not have python-pbr available. You have to use the Ubuntu Cloud Archive (Havana and up) repository to have it. This means they would have to backport python-pbr to Ubuntu Precise if they want to update python-jenkins to the latest.

It's not a strong -1. I'm however warning about the work this could generate for downstream packaging python-jenkins which is a detail we often forget.

Patch-set: 10