Update patch set 1
Patch Set 1:

It's certainly a possibility, which is why I've switched to trying to get some support in JJB ready first and I'll then experiment to see whether it can work with jenkinsapi.

Certainly there is a question long term as to whether it would be better to try and collaborate on one python wrapper for jenkins.

Patch-set: 1