Update patch set 3
Patch Set 3: Workflow-1

(4 comments)

Bit of work to be done here I thinks. Thanks for the feedback so far, query on whether the behaviour of get_plugins_info should be changed or retained to provide access to the raw info with the plugins property treated differently and intended for accessing the processed data.

Patch-set: 3
Label: Workflow=-1