Update patch set 3
Patch Set 3:

ahh. i thought it was needed for passing the params to jenkins.  Its hard to tell which URL requires a POST it's not really documented in the jenkins api doc (http://host:8080/api/), the only one it documents is the creatItem requests.  I'm wondering if the methods to create node, reconfigure job, reconfigure node requires it as well?

Patch-set: 3