Update patch set 3
Patch Set 3:

Well, only those that currently have a second input to the Request are providing data, and when doing so, that triggers Request to construct a POST request instead of a GET request. The only problem is that on python 3, these need to be byte strings, where as on python 2 they can be either.

Patch-set: 3