Update patch set 3
Patch Set 3:

(3 comments)

I think the only remaining work here is to ensure that user specified data is always a byte string or converted to a byte string before transmitting.

It's probably out of scope to determine which API calls should use POST requests and which shouldn't, however with this change, the Request object will build the request fully so if it should be a POST request it will ensure that the required data needed to trigger use of POST won't result in an exception being thrown due to use of str instead of unicode on python 3.

Patch-set: 3