Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

I think that the additional b'' are unnecessary for some of the entries as the Request object already has a second argument which just needs to be unicode() so adding .encode('utf-8') is sufficient in those cases.

Also it appears that for some of the calls b'' is being added as a second argument to jenkins_open() which is used to control add_crumb.

Seems like we could do with some stricter argument checking in some of the mocked objects to help catch these sort of issues.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1