Update patch set 6
Patch Set 6:

(4 comments)

Without the recursive list of jobs being returned it makes any use of folders with this library very difficult.

They may create a new API in Jenkins, but that will mean we need to provide a new function to retrieve that, and at the very least a mechanism to control which folder the current call is run against without needing to override the server.

Patch-set: 6