Update patch set 8
Patch Set 8: Code-Review-1

(4 comments)

Looking good in general. I think my main concern is that there needs to be some documentation for this since it will allow users to express folders in job names.   The other that's also needed for this change is the ability to get all jobs in get_jobs() method which was pointed out in PS 6.

@Jerome, you can just take what Darragh already has in his change and paste it into this one.  Let us know if you need help with that.

Patch-set: 8
Label: Code-Review=-1