Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I think the typical use case is to add/remove jobs from views.  Would be nice to have methods for that but I think can be done in a follow on change.  Just need to fix up the nit and i think this is ready.

Patch-set: 2
Label: Code-Review=-1