Update patch set 11
Patch Set 11:

(1 comment)

Only what is in the url templates at the top are used from the result of the function referenced. They may still be used to extract the folder name and job name, the url encoded ones are no longer used.

So might as well remove them from being encoded.

Patch-set: 11