Update patch set 9
Patch Set 9:

I agree with that, would be kinda inconsistent however I think get_plugin* was following the pattern set by get_job* which was probably a bad idea.  I think if we want to start over then a single get_plugins method is the correct thing to do.

Patch-set: 9