Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I'd like to see a little more English text in the examples, just to provide a bit more explanation of what the particular example is demonstrating.  Plugins in particular is probably rather…confusing, so a bit more about the expected output would probably be useful.  That said, the -1 is just for the marked inline issues…

Patch-set: 1
Label: Code-Review=-1