Update patch set 2
Patch Set 2:

(9 comments)

Looks like I missed a few things here and there. Some questions on whether we should strip out some of the exception catching currently going on.

Also may need to tweak the response_handler to make it clearer as to how to override in the future.

I believe that it's now safe to call ".decode('utf-8')" as the response.text is supposed to return unicode after converting correctly using chardet based on the headers and encoding specified. Although it might not hurt to test that bit out a little further.

Patch-set: 2
Label: Verified=0