Update patch set 2
Patch Set 2:

Probably an issue with how the NullServer is configured for the socket tests. If for some reason the connection attempt is rejected and the test returns immediately then the timeout will never fire.

I think the first step is an easy way to capture the output a little better. I suspect the trick to doing so may be to alter the method doing the time_limit for the timeout exception wraps the function it's calling to capture and output any exceptions before raising them. That'd make it easier to see what's happening.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0