Update patch set 21
Patch Set 21: Workflow-1

(3 comments)

Uploading where I am so far with a simple test that expose some bugs around the recursive get jobs code.

I do have a slight concern that we should consider adding a lightweight wrapper class to allow retrieval of the job fullname to distinguish this as something provided by this API for convenience versus what would be returned by jenkins.

One alternative may be to remove fullname for now, and add it back along with a lightweight wrapper class. Though that will make it slightly more awkward at the moment for applications using this.

Also another question would be, should we default the get_jobs to not be recursive or even have that as a get_all_jobs() method, and then have get_jobs call that with restricted to a single level?

Marking as WIP in progress so I can add the remaining tests over the weekend.

Sorry for taking so long to come up with these questions, just realised some of them when adding the new test.

Patch-set: 21
Label: Workflow=-1