Update patch set 5
Patch Set 5:

I think I'll leave consolidation of combining the tests for BadStatus, HTTPError, empty response and invalid json for a subsequent patch. It should be possible to define these as tests on the base class to be run for any of the real test classes provided I can think of a simple way to allow specify the api method to be called depending on the class.

Might need to be either a metaclass or dynamic method creation at class creation. But this would remove a lot of duplicate code.

Of course another alternative is to wait until the response handler and migration to requests has been added and only test those common exception behaviour via the response_handler instead of on each api call.

Patch-set: 5
Label: Verified=0