Update patch set 22
Patch Set 22:

still need to make that switch to get_all_jobs where get_jobs calls get_all_jobs with a depth restriction.

Main concern is just that we're adding to the dict of information coming back, so it's not something that is provided by Jenkins itself but it would look to consumers as though it was. Which can be seen from the need to update one of the tests to deal with the additional attribute in the returned data.

Patch-set: 22
Label: Verified=0