Update patch set 6
Patch Set 6:

So it appears that for post requests that they need to pass the query string parameters as a dictionary to the data parameters in the requests object. Not sure why this is necessary, but it means I'll have to test each api call to confirm correct behavior with requests and then alter the tests to correspond.

Patch-set: 6