Update patch set 3
Patch Set 3:

I'm sorry, I'm wrong; I was under the impression that we didn't have to install the plugin in our 1.624 deployment, but it turns out that it's there, so it may have just been a plugin that's installed by default.  What would you like me to do to the change for it to be acceptable to you?

Patch-set: 3
Label: Verified=0