Update patch set 2
Patch Set 2:

The intent of 35a5938d0f179942e52cf099b6eda3057702f239 as I read it was to consolidate url construction within python-jenkins so that every function did not have to be careful (urls are tricky).  Pushing the responsibility for url construction downstream to jjb and other libraries does not feel like a net benefit for them.

I also can't think of a use case where telling python-jenkins the url of the jenkins server and then it construct urls that do not point at the jenkins server would be intentional.

Patch-set: 2
Label: Verified=0