Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

Since this is really a private method, it's really dependent on the format_spec and job folder as to whether it starts with a '/'.

So it should never be called directly by an application, which allows some liberties to be taken.

That said, it's not impossible someone couldn't make a mistake with a future patch with regards to the template strings being used that wouldn't result in a leading slash being possible.

However you should use lstrip('/') to guard against it instead of assuming it's a single '/'.

Patch-set: 2
Label: Code-Review=-1