Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I still think this isn't actually getting to the bottom of what is going on.

The test is exercising an internal helper function, which on first glance cannot cause this behaviour.

I believe the correct thing to do here is to modify the tests to ensure we can exercise multiple variations of the base server to ensure each API call is behaviour as expected no matter what the server url is. Otherwise I think we'll just end up having this pop up again.


I've started this, and it showed that it's actually an internal usage of it that is broken.

Problem is in:

 def get_info(self, item="", query=None)

It constructs a url using '/'.join((item, INFO)), which when item is "" (default), will result in '/api/json' being passed to the _build_url method.


So it looks like the correct fix is to modify that call, and also add tests that exercise the entire API with both http://example.com and http://example.com/jenkins as the base server url.

Patch-set: 3
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1