Update patch set 3
Patch Set 3:

I'm not actively working on this patch.  It fixes the problem, but I could not get all green due to differences of opinion on the right approach.

Darragh started working on the fancy-test-harness path in https://review.openstack.org/#/c/231109/

Internally I've just kept the version pinned to something older, but am unsure how much longer that will be tenable.

Patch-set: 3