Update patch set 4
Patch Set 4: Code-Review-1

(3 comments)

Sorry, I don't know much about kerberos and how it works with Jenkins.  Was wondering if it requires installing a Jenkins plugin for this to work?  Also some documentation about this feature would be helpful.   Maybe add it to one of the files in under doc/source?  Tell that it's available, an overview, and provide references to required plugins and other related documentation?

Patch-set: 4
Label: Code-Review=-1