Update patch set 4
Patch Set 4: Code-Review-1

(6 comments)

You may wish to review the existing test suite, which you can find in tests in the top level.  We use "tox" to invoke tests in a virtual environment, so you may want to consider adding a "kerberos" dependency to the test-requirements.txt as well.

Patch-set: 4
Label: Code-Review=-1