Update patch set 4
Patch Set 4:

It should be possible—I'm only looking for unit tests.

If you're not familiar with the concept of mocking, you might want to take a look at the "mock" package, which is used extensively throughout the test suite.  The basic concept is that you take an external function your code calls and create a fake that returns some fixed value, then monkeypatch so your fake is called from that code instead.  You run your code, then you can optionally verify that the fake was called with the desired arguments.  The monkeypatching would require the "kerberos" package to be installed, but if done properly, nothing in the package would ever actually be called.

Patch-set: 4