Update patch set 1
Patch Set 1:

Could some reviewer help me understand why this does not work on Python 2.7 in your unit tests?

I have only run this on Python 2.7 myself, and it worked fine. The error I am getting:

Running function 'jenkins_open' resulted in exception 'TypeError' with message: 'urlopen() got an unexpected keyword argument 'context''

Indicates as if the wrong urlopen is used? I thought the idea was that the six library is supposed to make the API work the same for Py27 and Py3x? So it should be using urllib2.urlopen... Which from the docs has this context argument:
https://docs.python.org/2/library/urllib2.html#urllib2.urlopen

The other similar patches as seen in "Conflicts with" to the right on this page, seem to get the same issue.

Being able to pass your own SSL context here would be nice. So that it is possible to ignore verification of selfsigned SSL certs for testing.

Patch-set: 1