Update patch set 9
Patch Set 9: Workflow-1

I see that kerberos support has been added and unfortunately I'm not quite sure how to test such a configuration so I could properly implement support in the response handler to allow for doing the same.

For now I've just disabled it and marking this as a WIP.

Would appreciate some help in working out hot to do this.

Patch-set: 9
Label: Workflow=-1