Update patch set 9
Patch Set 9:

Darragh - not sure if you have in mind mock tests or testing in the complete kerberos env. If the latter - that would be total PITA to setup (really), we could rely on mock tests for regression and I could test your branch manually in our kerberos environment.

Patch-set: 9