Update patch set 1
Patch Set 1:

> @Joakim.  I'm seeing the same thing.  I'll try to dig a little more
 > to see what's going on.

Ok.  The reason why the test is failing is because the 'gate-python-jenkins-python27' test job is run with python 2.7.6.  I get the exact same error when I switch to run with the same python version.  It looks like the context parameter for urlopen was only added in python 2.7.9 and later. 

Additional reference:
 https://stackoverflow.com/questions/28575070/urllib-not-taking-context-as-a-parameter

Patch-set: 1