Update patch set 2
Patch Set 2: Code-Review+2

That is an aftermath of "Adds Cloudbees folder plugin support" https://review.openstack.org/#/c/180185/ 

The re.search() acts solely on the basename regardless of the job being in a folder or not. Based on the bug report that part works properly. Given jobs:

 job1
 folder/job2

A search for 'job' matches both.


The bug is later and the launchpad report hints about "jenkins.NotFoundException: Requested item could not be found".  That is triggered by get_job_info().

With the above example, we pass to get_job_info() ['job1', 'job2'].  And 'job2' is not going to be found since it does not exist.   By passing the fullname 'folder/job2', get_job_info() will find the job and append its information.

I try writing a test but gave up. I dont think it is needed in this case.

Patch-set: 2
Reviewer: Gerrit User 2475 <2475@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2