Update patch set 17
Patch Set 17:

ok, think I've solved the main issues, next step is to try and test the kerberos side. Don't have a setup that matched the original but as I'm using kerberos to auth with my employers AD, I might be able to get Jenkins to hook into it locally and then verify.

Patch-set: 17
Label: Verified=0