Update patch set 3
Patch Set 3:

(3 comments)

Recommend not mixing in multiple changes here.

I've been trying to work out how applications will distinguish between folders returned in such a case so as to be able to ensure they create them first. It may be that we just need to test whether the 'jobs' attribute is exposed in the returned data.

If it is, then I would go with the previous iteration of this patch and just return everything and let callers handle how to manage custom job types correctly.

Others might feel it's ok to put in custom handling so I wouldn't block a patch doing that, but it should be kept separate to this change.

Patch-set: 3
Label: Verified=0