Update patch set 21
Patch Set 21:

Hi Ken, at ONS last week we discussed with the OpenStack infra folks about changing the committers for this project as the current committers are not that active as of late. We got approval to promote new committers from the JJB team. I'm still travelling and won't be back until later this week but we plan to sort out the committers team and hopefully have some more active committers moving forward. Let me sync up with the new committers this week and see what we can do to get more activity in this project. Sorry for the long wait.

Patch-set: 21