Update patch set 4
Patch Set 4:

> Uploaded patch set 4.

Hi folks, I am not sure why my test fails on Zuul and not on my local. Seems to me that for some reason build_job_url is returning the parameters inside the url in different order every time and failing the assertion in the test. Any ideas on how deal with it, please?
Thanks,
Akshat

Patch-set: 4
Label: Verified=0