Update patch set 3
Patch Set 3:

(3 comments)

So I think we can simplify this a little and leave it up to requests to work out the correct think to do based on env variables provided we limit setting of the 'verify' attribute to only be modified on the self._session object only when we need to change it's value.

Not marking this as a -1 because I'm not 100% sure of the behaviour, this is just based on inferring from https://github.com/requests/requests/blob/d2e1130603c67d278eb49980205a3797e9e8b4a3/requests/sessions.py#L602 and not from validating this.

Patch-set: 3
Label: Workflow=0