Update patch set 1
Abandoned

This change orginated before we moved to use of python requests library which makes it mostly useless now.

With requests user can tune retries using https://stackoverflow.com/a/15431343/99834

Feel free to reopen it if you update it for requests and if it adds something that cannot be obtained using the mentioned workaround.

Patch-set: 1
Status: abandoned