Update patch set 2
Patch Set 2:

(5 comments)

You were missing the context of the fact that this was code from another project that worked fully, brought into here wholesale (I was monkey-patching python-jerkins to work).  So that is why the style/libraries used are different.

That said, I can make changes to bring it in line with the other tests.

Patch-set: 2