Update patch set 1
Patch Set 1:

(2 comments)

So, when using the REST API of Jenkins for api endpoints that can return a lot of data, if you do not specify depth or tree filters, Jenkins will respond with HTTP response 418 I’m a Teapot from RFC 2324.

So this patch adds the depth filter as a default to those api endpoints that will otherwise not work.

I can see how without this knowledge my commit message might not be fully clear, so to make sure it’s reworded in a way that you would be happy with, could you suggest what it should be please?

Patch-set: 1
Label: Verified=0