Update patch set 3
Patch Set 3:

Sorin, your patch don't fix my issue. Sorry for misleading, there is no issue with redirecting.
When build request is posted to Jenkins, there is HTTP status code 201 (Created) is in the response, and the header 'Location' provides the link to the queue where the job is waiting to start building.

My patch allows to get this header ('Location') if there is no 'content-lenght' in the response for the build request.

Patch-set: 3
Label: Verified=0