Update patch set 2
Patch Set 2:

(3 comments)

the commit message is a little tricky to read due to the length of a single line, and not necessarily the easiest to parse. Could you tweak that a little based on the comments I made so that hopefully someone looking back will find it easier to understand why the change was made.

Also might be clearer if the newline is added between the two scripts instead of part of the second script, but that's just me being cautious on helping future maintainers spot the importance of it to separate the scripts in case another change is needed.

Patch-set: 2
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0