Update patch set 4
Patch Set 4: Code-Review+1

(4 comments)

Looks good to me, I think some of the information in the commit message should be added to the docstring for the method, mainly just the explanation about including empty objects so that the code can guess whether to query again.

Could of other items I'm torn about or don't feel they are too important but may be worth considering. Keeping the fields definition at the top of the file in a library constant to allow overriding as desired might be useful, but unlikely to be used much. And the naming of the variable seems less descriptive than it could be, it's not directly related to folder_depth, though we could probably limit the value to something like min(folder_depth, 10) if folder_depth > 0, it's more how many levels of folders to include per request.

But these are mostly nits, so if others disagree, I'm ok for it to merge as is.

Patch-set: 4
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1