Update patch set 2
Patch Set 2:

The use case is that I am using Jenkins Job Builder to deploy jobs to a server. However that server is behind a corporate SSO proxy and the only way get JJB to be authenticated in such scenario is do a login externally, get a SSO token and then set that in a header for all subsequent HTTP requests to Jenkins. If that is in API, then all tools using python-jenkins would need separate support for this (rare) scenario

Patch-set: 2