Update patch set 7
Patch Set 7: Code-Review+1

(1 comment)

I'm kind of ok with this, but still feels like there should be a way as part of the API to set the same behaviour instead of making it environment variable only.

The reason the maybe_add_crumb doesn't need a log message to alert users is that it's entirely Jenkins behaviour controlled, the additional headers though would benefit because it's not a setting in Jenkins that turns them on/off, but rather the current executed environment, so someone looking at this being processed by a remote job output may not connect that someone set an env var on the machine the job is executed that affects it's behaviour.

So ok to land, but seems incomplete and likely someone will want to come back and emit a single log message (warning) see how we do the timeout warning for a way to emit a single warning https://git.openstack.org/cgit/openstack/python-jenkins/tree/jenkins/__init__.py?id=7c1a17c3f11eebd8ff4867d7ef1a7664fd3741d9#n287

But as I just realised that it this could actually do everything needed via the __init__ and the self._session object it should be possible to emit a single warning from the __init__ about picking up headers from the env variable.

Patch-set: 7
Label: Code-Review=+1