Update patch set 10
Patch Set 10: Code-Review-2

I am adding blocking this because an empty response without HTTP response code is in breach of HTTP spec and treating it as a success would be a big mistake.

I know for sure that some reverse proxies can in some conditions produce such behaviors and we do not want to ever consider a call succesful if we are not sure about that.

Is better to retry it until it works than assuming a success.

If you find more information like this being a documented Jenkins API behavior, I am willing to change my vote.

PS. -2 vote is only because I want to be sure that we don't introduce random failures. Maybe that is the time when we need to start testing with real jenkins containers as they woudl provide a much more realistic experience than the unittests.

Patch-set: 10