Update patch set 1
Patch Set 1: Code-Review-1

Hi Marcos, thanks for your contribution. I think this patch is trying to get too many ideas into 1 patch. Is it possible for you to split the patch into 3 separate patches for each of the 3 new additions so that we can review and comment on them as separate new features?

It would probably be good to add tests for the new functions in the tests directory as well so that we can ensure the feature continues to work as expected in the future.

Patch-set: 1
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1