Allow build number to be a string
When retreiving data from builds, the build number can have a string portion.
This is for matrix jobs that define a label to retrieve one sub-build of the
matrix job.

The URL structure is:
https://<jenkins_server>/job/>job_name>/<number>/label=<configuration>/

Allowing a string for build number allows for it to be specifed as, e.g.:
1/label=amd64

And whilst being somewhat a workaround, works without introducing complexity
and modifying funciton calls for extra parameters.

Change-Id: Ic409df6e20e85d6a417c6490472b3729f1f20cd2