Replace pkg_resources with packaging module
This fixes a warning in newer Python versions (>=3.10):
`DeprecationWarning: pkg_resources is deprecated as an API`.

pkg_resources is not part of the standard library. It is provided by
setuptools: https://github.com/pypa/setuptools/tree/main/pkg_resources.
This commit adds an explicit dependency on a third-party package
(packaging) that provides the same functionality.

Change-Id: I10b7db878cc6c346631497c41d9397d8b3379029